Add staff workflow endpoints & market prices
Introduce supermarket staff product workflow and market-price support: adds CustomersController (addresses, available stocklots), new ProductsController workflow endpoints for staff (identify, analyze-image with AI timeout, create product, create lot & publish) plus helper methods to resolve staff supermarket and display name. Adds OrdersController endpoints for delivery time-slots and collection points. Removes price feedback / accuracy endpoints from MarketPricesController. Update BarcodeLookupService to use ProductState enum and adjust product status handling.

Project and infra updates: add MSBuild target to clear static web asset cache, trim custom pipeline/swagger/cors/auth behaviors, and simplify Program startup pipeline call. Include multiple domain/entity/enum additions and DB migrations (market price/pricing history repairs, enum realignment, refund entity, delivery logs/legacy cleanup, coordinate fields, etc.) and related repository/service changes to support the new market-price and workflow features.

+++ src/CloseExpAISolution.Application/Services/Class/FeedbackService.cs
@@ -84,7 +84,7 @@ public class FeedbackService : IFeedbackService
             CustomerFeedbackId = Guid.NewGuid(),
             UserId = userId,
             OrderId = request.OrderId,
-            Rating = request.Rating,
+            Rating = (short)request.Rating,
             Comment = request.Comment,
             CreatedAt = DateTime.UtcNow
         };
@@ -106,7 +106,7 @@ public class FeedbackService : IFeedbackService
             return Error("Bạn không có quyền sửa đánh giá này");
 
         if (request.Rating.HasValue)
-            feedback.Rating = request.Rating.Value;
+            feedback.Rating = (short)request.Rating.Value;
 
         if (request.Comment != null)
             feedback.Comment = request.Comment;
