+++ src/CloseExpAISolution.Domain/Entities/CustomerFeedback.cs
@@ -5,11 +5,10 @@ public class CustomerFeedback
     public Guid CustomerFeedbackId { get; set; }
     public Guid UserId { get; set; }
     public Guid OrderId { get; set; }
-    public int Rating { get; set; }
+    public short Rating { get; set; }
     public string? Comment { get; set; }
     public DateTime CreatedAt { get; set; }
 
     public User? User { get; set; }
     public Order? Order { get; set; }
 }
-
